+++ src/Services/Dominion/Actions/WonderActionService.php
@@ -280,17 +280,18 @@ class WonderActionService
 
                 $wizardRatio = min(1, $this->militaryCalculator->getWizardRatioRaw($dominion));
                 $damageDealt = round($spellInfo['damage_multiplier'] * $wizardRatio * $this->landCalculator->getTotalLand($dominion));
+                $damageCap = static::CYCLONE_DAMAGE_CAP_PERCENTAGE / 100;
 
                 // Techs
                 $damageDealt *= (1 + $dominion->getTechPerkMultiplier('cyclone_damage'));
+                $damageCap *= (1 + $dominion->getTechPerkMultiplier('cyclone_damage'));
 
                 // Wonders
                 if ($wonder->wonder->perks->pluck('key')->contains('enemy_spell_damage')) {
                     $damageDealt *= (1 + $wonder->wonder->perks->groupBy('key')['enemy_spell_damage']->first()->pivot->value / 100);
                 }
 
-                // Cap at 2% of wonder max power
-                $damageCap = static::CYCLONE_DAMAGE_CAP_PERCENTAGE / 100;
+                // Cap at % of wonder max power
                 $damageDealt = min($damageDealt, round($wonder->power * $damageCap));
                 $dominion->stat_cyclone_damage += $damageDealt;